New stuff plus fixed
Added bunch of extra components and props to them, while also enabled support for AuthorAvatar for any iconUrl, thumbnail or Image prop.

Also enabled AuthorUsername in embed but planning to add it to every string prop too, so you can use it in place of embed author name or embed footer text or in button responses, etc.